requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
python-dotenv>=1.0.0
pyyaml>=6.0
psycopg[binary]>=3.1.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
